windows wip: read description
This almost works, but not quite. We're able to build the standard
library, builder, and test binaries, but each test fails to os.Chdir
into the test directory. Bazel works differently on Windows than on
other operating systems.

If I want to do a detailed article on Windows differences, this might
be worth explaining. Otherwise, it would really complicate the
examples.